fix a bug in CheckBlockHeader that was introduced with the AuxPow changes.
We were doing checkpoint checks in CheckBlockHeader, this caused the wallet
to flag the blockchain as corrupt if the last checkpoint is within -checkblocks
(default is 288) blocks of what is already stored in the local blockchain.

The checks are completed in ProcessBlock - this is where it was done in 0.9.2.15 and
prior, and bitcoin 0.9.2 upwards does the same.

see https://github.com/bitcoin/bitcoin/issues/4246 and resulting commit
https://github.com/sipa/bitcoin/commit/7a9e0b6460037b72376af293dc87acafe4fca831
for similar issue found during bitcoin 0.10 development.